Move DropWaker to its own module
