+++ src/lib.rs
@@ -147,6 +147,7 @@ use std::{fmt, ptr};
 
 mod bitmap;
 mod config;
+mod drop_waker;
 mod op;
 mod sys;
 
@@ -167,6 +168,7 @@ use bitmap::AtomicBitMap;
 pub use cancel::Cancel;
 use config::munmap;
 pub use config::Config;
+use drop_waker::drop_task_waker;
 #[doc(no_inline)]
 pub use extract::Extract;
 #[doc(no_inline)]
@@ -1010,60 +1012,6 @@ impl fmt::Debug for QueueFull {
 impl fmt::Display for QueueFull {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str("`a10::Ring` submission queue is full")
-    }
-}
-
-/// Create a [`task::Waker`] that will drop itself when the waker is dropped.
-///
-/// # Safety
-///
-/// The returned `task::Waker` cannot be cloned, it will panic.
-unsafe fn drop_task_waker<T: DropWaker>(to_drop: T) -> task::Waker {
-    unsafe fn drop_by_ptr<T: DropWaker>(data: *const ()) {
-        T::drop_from_waker_data(data);
-    }
-
-    // SAFETY: we meet the `task::Waker` and `task::RawWaker` requirements.
-    unsafe {
-        task::Waker::from_raw(task::RawWaker::new(
-            to_drop.into_waker_data(),
-            &task::RawWakerVTable::new(
-                |_| panic!("attempted to clone `a10::drop_task_waker`"),
-                // SAFETY: `wake` takes ownership, so dropping is safe.
-                drop_by_ptr::<T>,
-                |_| { /* `wake_by_ref` is a no-op. */ },
-                drop_by_ptr::<T>,
-            ),
-        ))
-    }
-}
-
-/// Trait used by [`drop_task_waker`].
-trait DropWaker {
-    /// Return itself as waker data.
-    fn into_waker_data(self) -> *const ();
-
-    /// Drop the waker `data` created by `into_waker_data`.
-    unsafe fn drop_from_waker_data(data: *const ());
-}
-
-impl<T> DropWaker for Box<T> {
-    fn into_waker_data(self) -> *const () {
-        Box::into_raw(self).cast()
-    }
-
-    unsafe fn drop_from_waker_data(data: *const ()) {
-        drop(Box::<T>::from_raw(data.cast_mut().cast()));
-    }
-}
-
-impl<T> DropWaker for Arc<T> {
-    fn into_waker_data(self) -> *const () {
-        Arc::into_raw(self).cast()
-    }
-
-    unsafe fn drop_from_waker_data(data: *const ()) {
-        drop(Arc::<T>::from_raw(data.cast_mut().cast()));
     }
 }
